fix: add support for long integers

+++ packages/openapi-ts/test/__snapshots__/2.0.x/plugins/zod/default/zod.gen.ts
@@ -2,21 +2,21 @@
 
 import { z } from 'zod';
 
-export const zCommentWithBreaks = z.number();
+export const zCommentWithBreaks = z.number().int();
 
-export const zCommentWithBackticks = z.number();
+export const zCommentWithBackticks = z.number().int();
 
-export const zCommentWithBackticksAndQuotes = z.number();
+export const zCommentWithBackticksAndQuotes = z.number().int();
 
-export const zCommentWithSlashes = z.number();
+export const zCommentWithSlashes = z.number().int();
 
-export const zCommentWithExpressionPlaceholders = z.number();
+export const zCommentWithExpressionPlaceholders = z.number().int();
 
-export const zCommentWithQuotes = z.number();
+export const zCommentWithQuotes = z.number().int();
 
-export const zCommentWithReservedCharacters = z.number();
+export const zCommentWithReservedCharacters = z.number().int();
 
-export const zSimpleInteger = z.number();
+export const zSimpleInteger = z.number().int();
 
 export const zSimpleBoolean = z.boolean();
 
@@ -47,7 +47,7 @@ export const zEnumFromDescription = z.number();
 
 export const zEnumWithExtensions = z.unknown();
 
-export const zArrayWithNumbers = z.array(z.number());
+export const zArrayWithNumbers = z.array(z.number().int());
 
 export const zArrayWithBooleans = z.array(z.boolean());
 
@@ -79,7 +79,7 @@ export const zDictionaryWithProperties = z.object({});
 export const zDate = z.string();
 
 export const zModelWithInteger = z.object({
-    prop: z.number().optional()
+    prop: z.number().int().optional()
 });
 
 export const zModelWithBoolean = z.object({
@@ -124,7 +124,7 @@ export const zModelWithEnum = z.object({
 });
 
 export const zModelWithEnumFromDescription = z.object({
-    test: z.number().optional()
+    test: z.number().int().optional()
 });
 
 export const zModelWithNestedEnums = z.object({
@@ -135,7 +135,7 @@ export const zModelWithNestedEnums = z.object({
         'Warning',
         'Error'
     ])).optional(),
-    arrayWithDescription: z.array(z.number()).optional()
+    arrayWithDescription: z.array(z.number().int()).optional()
 });
 
 export const zModelWithReference = z.object({
@@ -150,7 +150,7 @@ export const zModelWithReference = z.object({
         default: z.string().optional(),
         try: z.string().optional(),
         '@namespace.string': z.string().readonly().optional(),
-        '@namespace.integer': z.number().readonly().optional()
+        '@namespace.integer': z.number().int().readonly().optional()
     }).optional()
 });
 
@@ -181,7 +181,7 @@ export const zModelWithProperties = z.object({
     default: z.string().optional(),
     try: z.string().optional(),
     '@namespace.string': z.string().readonly().optional(),
-    '@namespace.integer': z.number().readonly().optional()
+    '@namespace.integer': z.number().int().readonly().optional()
 });
 
 export const zModelWithNestedProperties = z.object({
@@ -236,18 +236,18 @@ export const zModelWithPattern = z.object({
 
 export const zParameterActivityParams = z.object({
     description: z.string().optional(),
-    graduate_id: z.number().optional(),
-    organization_id: z.number().optional(),
-    parent_activity: z.number().optional(),
-    post_id: z.number().optional()
+    graduate_id: z.number().int().optional(),
+    organization_id: z.number().int().optional(),
+    parent_activity: z.number().int().optional(),
+    post_id: z.number().int().optional()
 });
 
 export const zResponsePostActivityResponse = z.object({
     description: z.string().optional(),
-    graduate_id: z.number().optional(),
-    organization_id: z.number().optional(),
-    parent_activity_id: z.number().optional(),
-    post_id: z.number().optional()
+    graduate_id: z.number().int().optional(),
+    organization_id: z.number().int().optional(),
+    parent_activity_id: z.number().int().optional(),
+    post_id: z.number().int().optional()
 });
 
 export const zFailureFailure = z.object({
@@ -268,7 +268,7 @@ export const zCallWithDuplicateResponsesResponse = zModelWithString;
 export const zCallWithResponsesResponse = z.union([
     z.object({
         '@namespace.string': z.string().readonly().optional(),
-        '@namespace.integer': z.number().readonly().optional(),
+        '@namespace.integer': z.number().int().readonly().optional(),
         value: z.array(zModelWithString).readonly().optional()
     }),
     zModelThatExtends,
